requirements.txt
supabase==2.10.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1